Initial commit
This is the initial commit of dvddump.

This is a simple wrapper around dd_rescue for the dumping of DVD's. It
uses libdvdread to first open the DVD, this takes care of al the CSS
stuff and allows dd_rescue to then access the DVD without errors.

Usage is as simple as

    $ dvddump /dev/sr0 dvd.iso

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>